+++ CHANGES.rst
@@ -2,8 +2,29 @@ Changelog
 =========
 
 
-0.1 (unreleased)
-----------------
+0.1.3 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
+0.1.2 (2016-01-15)
+------------------
+
+- Fix permission check in related objects portlet.
+  [cekk]
+
+
+0.1.1 (2016-01-15)
+------------------
+
+- Fix object retrieving in portlets and related items.
+  Now expired items are not showed. (#108)
+  [cekk]
+
+
+0.1.0 (2016-01-05)
+------------------
 
 - Initial release.
   [daniele-andreotti]